requirements1.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
google==3.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
google_auth_oauthlib==1.2.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
pyparsing==3.2.0
uritemplate==4.1.1
djangorestframework==3.15.2
gunicorn==23.0.0 
packaging==24.2
whitenoise==6.8.2